Rename n to iteration-count, count to...
...cell-count, and fix iteration-count.

Previously, one less iteration than the iteration-count argument was
being run because cell-count is based off iteration-count, and this was
one way to add the implicit empty cells on the ends of the automaton,
but it wasn't very clean.

To fix this, in the cell-count calculation I've added one to the
iteration-count prior to doubling it, and removed the subtract one from
iteration-count in the draw-update loop condition.